refactor(config): Refactor config package into separate files
