+++ config/config.go
@@ -2,167 +2,19 @@ package config
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"strings"
 	"time"
 
 	logging "github.com/op/go-logging"
 	"github.com/urfave/cli"
-	git "gopkg.in/src-d/go-git.v4"
-	yaml "gopkg.in/yaml.v2"
 )
 
 var configLogger = logging.MustGetLogger("config")
 
-type configFileV1 struct {
-	// config version
-	Version int `yaml:"version"`
-
-	CLI     configFileCLIV1
-	Analyze configFileAnalyzeV1
-
-	// what file this config is stored in
-	filename string
-}
-
-type configFileCLIV1 struct {
-	// Upload configuration.
-	APIKey  string `yaml:"api_key,omitempty"`
-	Server  string `yaml:"server,omitempty"`
-	Project string `yaml:"project,omitempty"`
-	Locator string `yaml:"locator,omitempty"`
-}
-
-type configFileAnalyzeV1 struct {
-	Modules []ModuleConfig `yaml:"modules,omitempty"`
-}
-
-// ModuleConfig defines a config for a builder's entry point
-type ModuleConfig struct {
-	Name string `yaml:"name"`
-	Path string `yaml:"path"`
-	Type string `yaml:"type"` // this is later transformed to a ModuleType
-}
-
-func readConfigFile(path string) (string, configFileV1, error) {
-	var configFilePath string
-	var configFile configFileV1
-	var parseError error
-	if path != "" {
-		if _, err := os.Stat(path); err != nil {
-			return path, configFileV1{}, fmt.Errorf("invalid config file specified")
-		}
-		configFilePath = path
-	}
-
-	if _, err := os.Stat(".fossa.yml"); err == nil {
-		configFilePath = ".fossa.yml"
-	} else if _, err = os.Stat(".fossa.yaml"); err == nil {
-		configFilePath = ".fossa.yaml"
-	}
-
-	if configFilePath != "" {
-		configFile, parseError = ParseConfigFile(configFilePath)
-	} else {
-		configFile, parseError = setDefaultValues(configFileV1{})
-	}
-	return configFilePath, configFile, parseError
-}
-
-// WriteConfigFile writes a config state to yaml
-func WriteConfigFile(conf *CliConfig) error {
-	if conf.ConfigFilePath == "" {
-		conf.ConfigFilePath = ".fossa.yml"
-	}
-
-	writeConfig := configFileV1{
-		Version: 1,
-		CLI: configFileCLIV1{
-			APIKey:  conf.APIKey,
-			Server:  conf.Endpoint,
-			Project: conf.Project,
-		},
-		Analyze: configFileAnalyzeV1{
-			Modules: conf.Modules,
-		},
-	}
-
-	yamlConfig, err := yaml.Marshal(writeConfig)
-	if err != nil {
-		return err
-	}
-
-	configHeader := []byte(`# Generated by FOSSA CLI (https://github.com/fossas/fossa-cli)
-# Visit https://fossa.io to learn more
-`)
-
-	return ioutil.WriteFile(conf.ConfigFilePath, append(configHeader, yamlConfig...), 0777)
-}
-
-// ParseConfigFile parses a file and returns a config state
-func ParseConfigFile(filename string) (configFileV1, error) {
-	// Read configuration file.
-	var config configFileV1
-
-	bytes, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return config, err
-	}
-
-	err = yaml.Unmarshal(bytes, &config)
-	if err != nil {
-		return config, err
-	}
-
-	config, err = setDefaultValues(config)
-	if err != nil {
-		return config, err
-	}
-
-	return config, nil
-}
-
-func setDefaultValues(c configFileV1) (configFileV1, error) {
-	// Set config version
-	c.Version = 1
-
-	// Set default endpoint.
-	if c.CLI.Server == "" {
-		c.CLI.Server = os.Getenv("FOSSA_ENDPOINT")
-		if c.CLI.Server == "" {
-			c.CLI.Server = "https://app.fossa.io"
-		}
-	}
-
-	// Load API key from environment variable.
-	if c.CLI.APIKey == "" {
-		c.CLI.APIKey = os.Getenv("FOSSA_API_KEY")
-	}
-
-	// Infer default locator and project from `git`.
-	if c.CLI.Locator == "" {
-		// TODO: this needs to happen in the module directory, not the working
-		// directory
-		repo, err := git.PlainOpen(".")
-		if err == nil {
-			project := c.CLI.Project
-			if project == "" {
-				origin, err := repo.Remote("origin")
-				if err == nil && origin != nil {
-					project = origin.Config().URLs[0]
-					c.CLI.Project = project
-				}
-			}
-
-			revision, err := repo.Head()
-			if err == nil {
-				c.CLI.Locator = "git+" + project + "$" + revision.Hash().String()
-			}
-		}
-	}
-
-	return c, nil
+// DefaultConfig specifies the config for the default cmd
+type DefaultConfig struct {
+	Build bool
 }
 
 // AnalyzeConfig specifies the config for the analyze cmd
@@ -186,8 +38,8 @@ type UploadConfig struct {
 	Data string
 }
 
-// CliConfig specifies the config available to the cli
-type CliConfig struct {
+// CLIConfig specifies the config available to the cli
+type CLIConfig struct {
 	APIKey   string
 	Project  string
 	Revision string
@@ -220,11 +72,6 @@ func MakeLocator(project string, revision string) string {
 	return "git+" + noHTTPSPrefix + "$" + revision
 }
 
-// DefaultConfig specifies the config for the default cmd
-type DefaultConfig struct {
-	Build bool
-}
-
 func parseModulesFlag(moduleFlag string) ([]ModuleConfig, error) {
 	if moduleFlag == "" {
 		return []ModuleConfig{}, nil
@@ -247,14 +94,14 @@ func parseModulesFlag(moduleFlag string) ([]ModuleConfig, error) {
 	return config, nil
 }
 
-// New creates a CliConfig from cli context
-func New(c *cli.Context) (CliConfig, error) {
+// New creates a CLIConfig from cli context
+func New(c *cli.Context) (CLIConfig, error) {
 	modules, err := parseModulesFlag(c.String("modules"))
 	if err != nil {
-		return CliConfig{}, err
+		return CLIConfig{}, err
 	}
 
-	var config = CliConfig{
+	var config = CLIConfig{
 		APIKey:         c.String("api_key"),
 		Project:        c.String("project"),
 		Revision:       c.String("revision"),
@@ -289,7 +136,7 @@ func New(c *cli.Context) (CliConfig, error) {
 	configFilePath, configFile, err := readConfigFile(config.ConfigFilePath)
 	config.ConfigFilePath = configFilePath
 	if err != nil {
-		return CliConfig{}, err
+		return CLIConfig{}, err
 	}
 
 	var locatorSections []string
